Example/api/requirements.txt
Django==1.11.6
djangorestframework==3.7.7
psycopg2==2.7.3.2
requests==2.19.1
